+++ StrategyImplementation/src/main/java/domain/CoffeeProvider.java
@@ -1,4 +1,4 @@
-package providers;
+package domain;
 
 import lombok.Value;
 import productline.ProductionLine;
