+++ StrategyImplementation/src/main/java/productline/impl/BrazilianAromaProductionLine.java
@@ -1,5 +1,6 @@
-package productline;
+package productline.impl;
 
+import productline.ProductionLine;
 import productline.guidelines.ProductionGuidelines;
 import scanners.LabelScanner;
